extend add LocalSigningService interface
