+++ src/main/java/org/eclipse/tractusx/managedidentitywallets/signing/SigningService.java
@@ -36,5 +36,4 @@ public interface SigningService {
     void saveKey(WalletKey key);
     SigningServiceType getSupportedServiceType();
     SignerResult createPresentation(PresentationCreationConfig config);
-    void setKeyProvider(KeyProvider keyProvider);
 }
